refactoring login/logout frontend functions
