+++ src/main/java/edu/example/salvo/SalvoApplication.java
@@ -246,6 +246,7 @@ class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 			.antMatchers("/web/styles/games.css").permitAll()
 			.antMatchers("/web/scripts/games.js").permitAll()
 			.antMatchers("/web/scripts/handleTime.js").permitAll()
+			.antMatchers("/web/scripts/loginFunctions.js").permitAll()
 			.antMatchers("/api/games").permitAll()
 			.antMatchers("/api/leaderboard").permitAll()
 			.antMatchers("/api/players").permitAll()
